refactored calendar
