+++ app/components/Calendar/index.js
@@ -1,0 +1,5 @@
+import Header from './Header.js';
+import Calendar from './Calendar.js';
+import SliderEntry from './SliderEntry.js';
+
+export { Header, Calendar, SliderEntry };
