+++ app/components/Calendar/styles/SliderEntry.style.js
@@ -1,5 +1,4 @@
 import { StyleSheet, Dimensions, Platform } from 'react-native';
-import { colors } from './index.style';
 
 const IS_IOS = Platform.OS === 'ios';
 const { width: viewportWidth, height: viewportHeight } = Dimensions.get('window');
@@ -26,27 +25,12 @@ export default StyleSheet.create({
         paddingHorizontal: itemHorizontalMargin,
         paddingBottom: 18 // needed for shadow
     },
-    shadow: {
-        position: 'absolute',
-        top: 0,
-        left: itemHorizontalMargin,
-        right: itemHorizontalMargin,
-        bottom: 18,
-        shadowColor: colors.black,
-        shadowOpacity: 0.25,
-        shadowOffset: { width: 0, height: 10 },
-        shadowRadius: 10,
-        borderRadius: entryBorderRadius
-    },
     imageContainer: {
         flex: 1,
         marginBottom: IS_IOS ? 0 : -1, // Prevent a random Android rendering issue
         backgroundColor: 'white',
         borderTopLeftRadius: entryBorderRadius,
         borderTopRightRadius: entryBorderRadius
-    },
-    imageContainerEven: {
-        backgroundColor: colors.black
     },
     image: {
         ...StyleSheet.absoluteFillObject,
@@ -64,9 +48,6 @@ export default StyleSheet.create({
         height: entryBorderRadius,
         backgroundColor: 'white'
     },
-    radiusMaskEven: {
-        backgroundColor: colors.black
-    },
     textContainer: {
         justifyContent: 'center',
         paddingTop: 20 - entryBorderRadius,
@@ -76,23 +57,8 @@ export default StyleSheet.create({
         borderBottomLeftRadius: entryBorderRadius,
         borderBottomRightRadius: entryBorderRadius
     },
-    textContainerEven: {
-        backgroundColor: colors.black
-    },
-    title: {
-        color: colors.black,
-        fontSize: 13,
-        fontWeight: 'bold',
-        letterSpacing: 0.5
-    },
     titleEven: {
         color: 'white'
-    },
-    subtitle: {
-        marginTop: 6,
-        color: colors.gray,
-        fontSize: 12,
-        fontStyle: 'italic'
     },
     subtitleEven: {
         color: 'rgba(255, 255, 255, 0.7)'
