--- app/components/index.js
@@ -1,6 +1,0 @@
-import Header from './Header.js';
-import Calendar from './Calendar.js';
-import InfiniteCalendar from './InfiniteCalendar.js';
-import Agenda from './Agenda.js';
-
-export { Header, Calendar, InfiniteCalendar, Agenda };
